chore: move DTO classes to base package and update imports
