+++ backend/src/main/java/ca/bc/gov/nrs/hrs/dto/base/CodeDescriptionDto.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.nrs.hrs.dto;
+package ca.bc.gov.nrs.hrs.dto.base;
 
 import lombok.With;
