+++ backend/src/main/java/ca/bc/gov/nrs/hrs/security/RoleConstants.java
@@ -1,6 +1,6 @@
 package ca.bc.gov.nrs.hrs.security;
 
-import ca.bc.gov.nrs.hrs.dto.Role;
+import ca.bc.gov.nrs.hrs.dto.base.Role;
 import org.springframework.stereotype.Component;
